Update Sum of Two Numbers
Write a program that checks all possible combinations of a pair of numbers in the interval of two given numbers. The output is printed in which order is the combination whose sum of numbers is equal to a given magic number. If there is no combination matching the condition, a message is printed that it was not found.

Login:
The input is read from the console and consists of three lines:
• First row - start of interval - integer in the interval [1...999]
• Second row - end of range - integer in range [greater than first number...1000]
• Third line – the magic number – an integer in the interval [1...10000].

Output:
One line should be printed to the console, according to the result:
• If a combination is found whose sum of numbers is equal to the magic number, we print: "Combination N:{sequence number} ({first number} + {second number} = {magic number})".
• If no combination matching the condition is found, we print: "{the number of all combinations} combinations - neither equals {the magic number}".