Update Building
Write a program that displays the room numbers in a building (in descending order) on the console if the following conditions are met:

• There are only offices on each even-numbered floor.
• There are only apartments on each odd-numbered floor.

Each apartment is designated as follows: A{floor number}{apartment number}, apartment numbers start from 0.

Each office is designated as follows: O{floor number}{office number}, office numbers also start from 0.

There are always apartments on the top floor and they are larger than the others, that's why they have an 'L' in front of the number instead of an 'A'. If there is only one floor, then there are only large apartments!
Two integers are read from the console - the number of floors and the number of rooms per floor.