Create Equal Sums Even Odd Position
Write a program that reads from the console two six-digit integers in the range 100,000 to 300,000. The first number entered will always be less than the second. On the console, print on 1 line, separated by a space, all the numbers that are between the two numbers read by the console and meet the following condition:
The sum of the digits in even and odd positions must be equal. If there are no numbers matching the condition on the console, no result is displayed.