Update Sum Prime Non Prime
Write a program that reads from the console integers in the range until a "stop" command is received. To find the sum of all prime numbers entered and the sum of all non-prime numbers entered. Since, by definition in mathematics, negative numbers cannot be prime, if a negative number is given as an input, the following message "Number is negative." should be displayed. In this case, an entered number is ignored and not added to either sum, and the program continues it's execution, waiting for the next number to be entered.
On the output, print the two sums found on two lines in the following format:
• "Sum of all prime numbers is: {prime numbers sum}".
• "Sum of all non prime numbers is: {nonprime numbers sum}".