Update Special Numbers
Write a program that reads an integer N entered by the user and generates all possible "special" numbers from 1111 to 9999. 
For a number to be "special", it must satisfy the following condition:
• N to be divided by each of its digits without a remainder.
Example: for N = 16
• 16 / 1 = 16 with no remainder
• 16 / 2 = 8 with no remainder
• 16 / 4 = 4 with no remainder
• 16 / 8 = 2 with no remainder

Login:
The input is read from the console and consists of a single integer in the range [1…600000].

Exit:
All "special" numbers separated by a space should be printed on the console.